Extract files from Auth Service to Family Service

+++ src/Application/Services/FamilyService.cs
@@ -1,8 +1,14 @@
+using System.IdentityModel.Tokens.Jwt;
+using System.Security.Claims;
 using Application.Dto;
 using Application.Interfaces;
 using AutoMapper;
+using Domain.Consts;
 using Domain.Entities;
+using Domain.Exceptions;
 using Infrastructure.Interfaces;
+using Midas.Services;
+using Microsoft.AspNetCore.Http;
 
 namespace Application.Services;
 
@@ -10,16 +16,28 @@ public class FamilyService : IFamilyService
 {
     private readonly IFamilyRepository _familyRepository;
     private readonly IMapper _mapper;
+    private readonly IUserClient _userClient;
+    private readonly IHttpContextAccessor _httpContextAccessor;
 
-    public FamilyService(IFamilyRepository familyRepository, IMapper mapper)
+    public FamilyService(
+        IFamilyRepository familyRepository, IMapper mapper, IUserClient userClient, IHttpContextAccessor contextAccessor)
     {
         _familyRepository = familyRepository;
         _mapper = mapper;
+        _userClient = userClient;
+        _httpContextAccessor = contextAccessor;
     }
     
     public async Task<AddNewFamilyReturnDto> AddNewFamily(AddNewFamilyDto dto)
     {
         var familyEntity = _mapper.Map<AddNewFamilyDto, Family>(dto);
+        var userEntity = await _userClient.GetUserByIdAsync((long)dto.FounderId).ConfigureAwait(false);
+
+        if (userEntity is null)
+        {
+            throw new UserException("");
+        }
+
         var familyId = await _familyRepository.AddNewFamily(familyEntity);
 
         return new AddNewFamilyReturnDto
@@ -31,12 +49,17 @@ public class FamilyService : IFamilyService
 
     public async Task<bool> DeleteFamily(ulong id)
     {
-        return await _familyRepository.DeleteFamily(id).ConfigureAwait(false);
+        if (await CheckIfUserInFamilyHasRole(id, FamilyRoles.MainAdministrator))
+        {
+            return await _familyRepository.DeleteFamily(id).ConfigureAwait(false);
+        }
+
+        return false;
     }
 
     public async Task<bool> AddUserToFamily(AddUserToFamilyDto dto)
     {
-        var userEntity = await _userRepository.GetUserByEmail(dto.Email).ConfigureAwait(false);
+        var userEntity = await _userClient.GetUserByEmailAsync(dto.Email).ConfigureAwait(false);
         var familyEntity = await _familyRepository.GetFamilyById(dto.FamilyId).ConfigureAwait(false);
 
         if (userEntity is null || familyEntity is null)
@@ -44,12 +67,17 @@ public class FamilyService : IFamilyService
             return false;
         }
 
-        return await _familyRepository.AddUserToFamily(userEntity.Id, familyEntity.Id).ConfigureAwait(false);
+        return await _familyRepository.AddUserToFamily((ulong)userEntity.Id, familyEntity.Id).ConfigureAwait(false);
     }
 
     public async Task<bool> DeleteUserFromFamily(DeleteUserFromFamilyDto dto)
     {
-        var userEntity = await _userRepository.GetUserByEmail(dto.Email).ConfigureAwait(false);
+        if (!await CheckIfUserInFamilyHasRole(dto.FamilyId, FamilyRoles.MainAdministrator))
+        {
+            return false;
+        }
+        
+        var userEntity = await _userClient.GetUserByEmailAsync(dto.Email).ConfigureAwait(false);
         var familyEntity = await _familyRepository.GetFamilyById(dto.FamilyId).ConfigureAwait(false);
 
         if (userEntity is null || familyEntity is null)
@@ -57,12 +85,17 @@ public class FamilyService : IFamilyService
             return false;
         }
 
-        return await _familyRepository.DeleteUserFromFamily(userEntity.Id, familyEntity.Id).ConfigureAwait(false);
+        return await _familyRepository.DeleteUserFromFamily((ulong)userEntity.Id, familyEntity.Id).ConfigureAwait(false);
     }
 
     public async Task<bool> SetUserFamilyRole(SetUserFamilyRoleDto dto)
     {
-        var userEntity = await _userRepository.GetUserByEmail(dto.Email).ConfigureAwait(false);
+        if (!await CheckIfUserInFamilyHasRole(dto.FamilyId, FamilyRoles.MainAdministrator))
+        {
+            return false;
+        }
+        
+        var userEntity = await _userClient.GetUserByEmailAsync(dto.Email).ConfigureAwait(false);
         
         if (userEntity is null)
         {
@@ -70,8 +103,56 @@ public class FamilyService : IFamilyService
         }
         
         var userFamilyRole = _mapper.Map<SetUserFamilyRoleDto, UserFamilyRole>(dto, opt => 
-            opt.AfterMap((_, dest) => dest.UserId = userEntity.Id));
+            opt.AfterMap((_, dest) => dest.UserId = (ulong)userEntity.Id));
         
         return await _familyRepository.SetUserFamilyRole(userFamilyRole).ConfigureAwait(false);
+    }
+
+    public async Task<UserFamilyRoleListDto> GetFamilyMembersForActiveUser()
+    {
+        var token = _httpContextAccessor.HttpContext.Request.Headers["Authorization"].ToString();
+        var handler = new JwtSecurityTokenHandler();
+
+        var userId = handler.ReadJwtToken(token).Claims.Single(x => x.Type == ClaimTypes.NameIdentifier).Value;
+        var userFamilyRoles = await _familyRepository.GetFamilyMemberRolesForUser(Convert.ToUInt64(userId)).ConfigureAwait(false);
+        var userFamilyTasks = userFamilyRoles.Select(MapUserFamilyRoleToDto);
+        var userFamilyRoleDtos = await Task.WhenAll(userFamilyTasks);
+
+        return _mapper.Map<List<UserFamilyRoleDto>, UserFamilyRoleListDto>(userFamilyRoleDtos.ToList());
+    }
+
+    private async Task<UserFamilyRoleDto> MapUserFamilyRoleToDto(UserFamilyRole userFamilyRole)
+    {
+        var userEntity = await _userClient.GetUserByIdAsync((long)userFamilyRole.UserId);
+
+        if (userEntity is null) return null;
+
+        var familyRoleDto = _mapper.Map<FamilyRole, FamilyRoleDto>(userFamilyRole.FamilyRole);
+        var familyDto = _mapper.Map<Family, FamilyDto>(userFamilyRole.Family);
+
+        var userFamilyRoleDto = new UserFamilyRoleDto
+        {
+            User = userEntity,
+            FamilyRole = familyRoleDto,
+            Family = familyDto
+        };
+            
+        return userFamilyRoleDto;
+    }
+
+    private async Task<bool> CheckIfUserInFamilyHasRole(ulong familyId, FamilyRoles role)
+    {
+        var token = _httpContextAccessor.HttpContext.Request.Headers["Authorization"].ToString();
+        var handler = new JwtSecurityTokenHandler();
+        var userId = Convert.ToUInt64(handler.ReadJwtToken(token).Claims.Single(x => x.Type == ClaimTypes.NameIdentifier).Value);
+        var userFamilyRoles = await _familyRepository.GetFamilyMemberRolesForUser(userId).ConfigureAwait(false);
+        var userInFamilyRole = userFamilyRoles?.SingleOrDefault(x => x.FamilyId == familyId);
+
+        if (userInFamilyRole is null)
+        {
+            return false;
+        }
+        
+        return (int)userInFamilyRole.FamilyRoleId == (int)role;
     }
 }
